+++ lib/sinatra/swagger-exposer/swagger-parameter-preprocessor.rb
@@ -1,4 +1,5 @@
 require_relative 'swagger-endpoint-parameter'
+require_relative 'swagger-parameter-helper'
 require_relative 'swagger-invalid-exception'
 
 module Sinatra
@@ -7,6 +8,8 @@ module Sinatra
 
     # Process the parameters for validation and enrichment
     class SwaggerParameterPreprocessor
+
+      include SwaggerParameterHelper
 
       def initialize(name, how_to_pass, required, type, default, params)
         @name = name.to_s
@@ -17,22 +20,25 @@ module Sinatra
         @params = params
 
         # All headers are upcased
-        if how_to_pass == SwaggerEndpointParameter::HOW_TO_PASS_HEADER
+        if how_to_pass == HOW_TO_PASS_HEADER
           @name.upcase!
         end
       end
 
       def useful?
-        @required || (!@default.nil?) || [SwaggerEndpointParameter::TYPE_NUMBER, SwaggerEndpointParameter::TYPE_INTEGER, SwaggerEndpointParameter::TYPE_BOOLEAN].include?(@type)
+        @required ||
+            (!@default.nil?) ||
+            [TYPE_NUMBER, TYPE_INTEGER, TYPE_BOOLEAN].include?(@type) || # Must check type
+            (@params.key? PARAMS_MIN_LENGTH) || (@params.key? PARAMS_MAX_LENGTH) # Must check string
       end
 
       def run(app, parsed_body)
         case @how_to_pass
-          when SwaggerEndpointParameter::HOW_TO_PASS_QUERY, SwaggerEndpointParameter::HOW_TO_PASS_PATH
+          when HOW_TO_PASS_QUERY, HOW_TO_PASS_PATH
             check_param(app.params)
-          when SwaggerEndpointParameter::HOW_TO_PASS_HEADER
+          when HOW_TO_PASS_HEADER
             check_param(app.headers)
-          when SwaggerEndpointParameter::HOW_TO_PASS_BODY
+          when HOW_TO_PASS_BODY
             check_param(parsed_body || {})
         end
       end
@@ -50,17 +56,18 @@ module Sinatra
 
       def validate_param_value(value)
         case @type
-          when SwaggerEndpointParameter::TYPE_NUMBER
+          when TYPE_NUMBER
             return validate_param_value_number(value)
-          when SwaggerEndpointParameter::TYPE_INTEGER
+          when TYPE_INTEGER
             return validate_param_value_integer(value)
-          when SwaggerEndpointParameter::TYPE_BOOLEAN
+          when TYPE_BOOLEAN
             return validate_param_value_boolean(value)
           else
-            return value
+            return validate_param_value_string(value)
         end
       end
 
+      # Validate a boolean parameter
       def validate_param_value_boolean(value)
         if (value == 'true') || value.is_a?(TrueClass)
           return true
@@ -71,6 +78,7 @@ module Sinatra
         end
       end
 
+      # Validate an integer parameter
       def validate_param_value_integer(value)
         begin
           f = Float(value)
@@ -90,6 +98,7 @@ module Sinatra
         end
       end
 
+      # Validate a number parameter
       def validate_param_value_number(value)
         begin
           value = Float(value)
@@ -107,16 +116,38 @@ module Sinatra
       def validate_numerical_value(value)
         validate_numerical_value_internal(
             value,
-            SwaggerEndpointParameter::PARAMS_MINIMUM,
-            SwaggerEndpointParameter::PARAMS_EXCLUSIVE_MINIMUM,
+            PARAMS_MINIMUM,
+            PARAMS_EXCLUSIVE_MINIMUM,
             '>=',
             '>')
         validate_numerical_value_internal(
             value,
-            SwaggerEndpointParameter::PARAMS_MAXIMUM,
-            SwaggerEndpointParameter::PARAMS_EXCLUSIVE_MAXIMUM,
+            PARAMS_MAXIMUM,
+            PARAMS_EXCLUSIVE_MAXIMUM,
             '<=',
             '<')
+      end
+
+      # Validate a string parameter
+      def validate_param_value_string(value)
+        if value
+          validate_param_value_string_length(value, PARAMS_MIN_LENGTH, '>=')
+          validate_param_value_string_length(value, PARAMS_MAX_LENGTH, '<=')
+        end
+        value
+      end
+
+      # Validate the length of a string parameter
+      # @param value the value to check
+      # @param limit_param_name [Symbol] the param that contain the value to compare to
+      # @param limit_param_method [String] the comparison method to call
+      def validate_param_value_string_length(value, limit_param_name, limit_param_method)
+        if @params.key? limit_param_name
+          target_value = @params[limit_param_name]
+          unless value.length.send(limit_param_method, target_value)
+            raise SwaggerInvalidException.new("Parameter [#{@name}] length should be #{limit_param_method} than #{target_value} but is #{value.length} for [#{value}]")
+          end
+        end
       end
 
       # Validate the value of a number