+++ CHANGELOG.md
@@ -1,8 +1,13 @@
+# 0.2.0
+
+- Fix empty value returned from validation when a parameter is a String
+- Added maxLength and minLength validations
+
 # 0.1.0
 
 - Added verification and enrichment form params
 - Added type extends
-- Added validation for min and max value
+- Added validation for min and max value for numbers
 
 # 0.0.1